% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Number.R
\name{Add_Number}
\alias{Add_Number}
\title{Add a number.}
\usage{
Add_Number(i, n)
}
\arguments{
\item{i}{Number.}

\item{n}{Number to add.}
}
\value{
List.  Output a named list.  Element \code{success} is a boolean denoting if the function was successful.
    Element \code{output} contains the function result (a numeric in the case of a success) or an error message
    (in the case of a failure)
}
\description{
Node to add a number to the incoming number and return the sum
}
\examples{
\dontrun{
Add_Number(3, 5)
}

}

