% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_TwoNumbers.R
\name{Add_TwoNumbers}
\alias{Add_TwoNumbers}
\title{Add two numbers.}
\usage{
Add_TwoNumbers(num1, num2)
}
\arguments{
\item{num1}{Number.}

\item{num2}{Number.}
}
\value{
List.  Output a named list.  Element \code{success} is a boolean denoting if the function was successful.
    Element \code{output} contains the function result (a numeric in the case of a success) or an error message
    (in the case of a failure)
}
\description{
Node to add two numerical inputs together and return the sum
}
\examples{
\dontrun{
Add_TwoNumbers(3, 5)
}

}

