% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num_Initialize.R
\name{Num_Initialize}
\alias{Num_Initialize}
\title{Set initial number.}
\usage{
Num_Initialize(n)
}
\arguments{
\item{n}{Number to add.}

\item{i}{Number.}
}
\value{
List.  Output a named list.  Element \code{success} is a boolean denoting if the function was successful.
    Element \code{output} contains the function result (a numeric in the case of a success) or an error message
    (in the case of a failure)
}
\description{
Node to set an initial number
}
\examples{
\dontrun{
Num_Initialize(3)
}

}

