% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wait.R
\name{Wait}
\alias{Wait}
\title{Wait n seconds.}
\usage{
Wait(i, n)
}
\arguments{
\item{i}{Number (input).}

\item{n}{Time to wait (seconds).}
}
\value{
List.  Output a named list.  Element \code{success} is a boolean denoting if the function was successful.
    Element \code{output} contains the function result (the input i in the case of a success) or an error message
    (in the case of a failure)
}
\description{
Node to wait for testing purposes
}
\examples{
\dontrun{
Wait(3, 5)
}

}

