% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Number.R
\name{Add_Number}
\alias{Add_Number}
\title{Add a number.}
\usage{
Add_Number(i, n)
}
\arguments{
\item{i}{Number.}

\item{n}{Number to add.}
}
\value{
A Number.
}
\description{
Node to add a number to the incoming number and return the sum
}
\examples{
\dontrun{
Add_Number(3, 5)
}

}

