% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num_Initialize.R
\name{Num_Initialize}
\alias{Num_Initialize}
\title{Set initial number.}
\usage{
Num_Initialize(n)
}
\arguments{
\item{n}{Number to add.}

\item{i}{Number.}
}
\value{
A Number.
}
\description{
Node to set an initial number
}
\examples{
\dontrun{
Add_Number(3, 5)
}

}

