% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wait.R
\name{Wait}
\alias{Wait}
\title{Wait n seconds.}
\usage{
Wait(i, n)
}
\arguments{
\item{i}{Number (input).}

\item{n}{Time to wait (seconds).}
}
\value{
A Number.
}
\description{
Node to wait for testing purposes
}
\examples{
\dontrun{
Wait(3, 5)
}

}

